Move errors to a separate file
